32-bit fixes (#2397)
Reorganize p4c headers so that we can reuse bitops.h from p4c/lib
* use int64_t for long
* fix hash column initialization
* fix parsing of ints and big ints
* reorganized headers and remove warnings

Big integers are stored in vectors of uintptr_t and thus the comparison
should be bit-width dependent. bitvecs are represented at uintptr_t, so seeds that are larger
than 32 bits need to extract all the components of the vector.

* correctly assign to seeds larger than 32-bits on 32-bit systems
* missing #ifdefs and warnings on 64-bit linux (gcc 6.4.0) and macos
* update p4c refpoint

+++ bf-asm/json.cpp
@@ -17,7 +17,7 @@ std::istream &operator>>(std::istream &in, std::unique_ptr<obj> &json) {
             /* fall through */
         case '0': case '1': case '2': case '3': case '4':
         case '5': case '6': case '7': case '8': case '9': {
-            long long l = 0;
+            int64_t l = 0;
             while (in && isdigit(ch)) {
                 /* FIXME -- deal with overflow ... and hex? */
                 l = l * 10 + ch - '0';
